pyproject.toml
[tool.poetry]
name = "lgr-pb"
version = "0.1.0"
description = ""
authors = ["BlueGlassBlock <blueglassblock@outlook.com>"]
license = "AGPL-3.0"

[[tool.poetry.source]]
name = "tuna-tsinghua"
default = false
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"

[[tool.poetry.source]]
name = "sustech"
default = false
url = "https://mirrors.sustech.edu.cn/pypi/simple/"

[[tool.poetry.source]]
name = "aliyun"
default = false
url = "https://mirrors.aliyun.com/pypi/simple/"

[[tool.poetry.source]]
name = "tencent-cloud"
default = false
url = "https://mirrors.cloud.tencent.com/pypi/simple"

[[tool.poetry.source]]
name = "netease"
default = false
url = "https://mirrors.163.com/pypi/simple/"


[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^8.4.0"
aiohttp = "^3.8.1"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
